refactor(common)!: move column stats types under org.apache.hudi.metadata.stats
Relocates the column statistics vocabulary (HoodieColumnRangeMetadata,
ValueType, ValueMetadata, ParquetAdapter, and spark-client's
SparkValueMetadataUtils) from org.apache.hudi.stats to
org.apache.hudi.metadata.stats. These types are metadata-subsystem
vocabulary: they depend on metadata.HoodieIndexVersion, and
ValueType.ordinal() is persisted in HoodieMetadataColumnStats records
of the col-stats/partition-stats index. Nesting them under
org.apache.hudi.metadata - a namespace that is permanently anchored by
persisted FQNs (HoodieMetadataPayload in every metadata table's
hoodie.properties) - gives them a home that never moves again.

BREAKING CHANGE: source/binary break for externally compiled code
using these types programmatically (e.g. reading footer ranges via
FileFormatUtils APIs); fails loudly, fixed by an import change. No
config values, reflection targets, or Kryo registrations reference
these FQNs, and nothing persisted carries Java class names - metadata
records store typeOrdinal ints and commit metadata serializes by field
name. Note ValueType's enum ORDER is storage-format-locked via the
persisted ordinal (independent of this move).

Verified across common, hadoop-common, client-common, spark client and
datasource, utilities (spark4.1) and flink-client (flink2.1); value
type/metadata and metadata payload suites pass.
