+++ hudi-common/src/main/java/org/apache/hudi/common/model/HoodieWriteStat.java
@@ -20,7 +20,7 @@ package org.apache.hudi.common.model;
 
 import org.apache.hudi.common.table.read.HoodieReadStats;
 import org.apache.hudi.common.util.Option;
-import org.apache.hudi.stats.HoodieColumnRangeMetadata;
+import org.apache.hudi.metadata.stats.HoodieColumnRangeMetadata;
 import org.apache.hudi.storage.StoragePath;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
