+++ hudi-common/src/main/java/org/apache/hudi/common/util/HFileUtils.java
@@ -47,7 +47,7 @@ import org.apache.hudi.io.hfile.HFileWriterImpl;
 import org.apache.hudi.io.hfile.UTF8StringKey;
 import org.apache.hudi.keygen.BaseKeyGenerator;
 import org.apache.hudi.metadata.HoodieIndexVersion;
-import org.apache.hudi.stats.HoodieColumnRangeMetadata;
+import org.apache.hudi.metadata.stats.HoodieColumnRangeMetadata;
 import org.apache.hudi.storage.HoodieStorage;
 import org.apache.hudi.storage.StoragePath;
