+++ hudi-common/src/main/java/org/apache/hudi/core/io/storage/ColumnRangeMetadataProvider.java
@@ -19,7 +19,7 @@
 package org.apache.hudi.core.io.storage;
 
 import org.apache.hudi.metadata.HoodieIndexVersion;
-import org.apache.hudi.stats.HoodieColumnRangeMetadata;
+import org.apache.hudi.metadata.stats.HoodieColumnRangeMetadata;
 
 import java.util.Map;
