+++ hudi-common/src/main/java/org/apache/hudi/metadata/stats/ParquetAdapter.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.hudi.stats;
+package org.apache.hudi.metadata.stats;
 
 import org.apache.hudi.common.util.ReflectionUtils;
