+++ hudi-common/src/main/java/org/apache/parquet/schema/LogicalTypeParquetAdapter.java
@@ -19,8 +19,8 @@
 
 package org.apache.parquet.schema;
 
-import org.apache.hudi.stats.ParquetAdapter;
-import org.apache.hudi.stats.ValueType;
+import org.apache.hudi.metadata.stats.ParquetAdapter;
+import org.apache.hudi.metadata.stats.ValueType;
 
 /**
  * Uses LogicalTypeAnnotation to extract value type, precision, and scale
