+++ hudi-common/src/test/java/org/apache/hudi/common/model/TestHoodieWriteStat.java
@@ -23,8 +23,8 @@ import org.apache.hudi.common.table.HoodieTableConfig;
 import org.apache.hudi.common.table.timeline.TimelineUtils;
 import org.apache.hudi.common.util.Functions;
 import org.apache.hudi.common.util.collection.Pair;
-import org.apache.hudi.stats.HoodieColumnRangeMetadata;
-import org.apache.hudi.stats.ValueMetadata;
+import org.apache.hudi.metadata.stats.HoodieColumnRangeMetadata;
+import org.apache.hudi.metadata.stats.ValueMetadata;
 import org.apache.hudi.storage.StoragePath;
 
 import org.apache.avro.util.Utf8;
