+++ hudi-common/src/test/java/org/apache/hudi/common/util/TestBaseFileUtils.java
@@ -20,8 +20,8 @@
 package org.apache.hudi.common.util;
 
 import org.apache.hudi.metadata.HoodieIndexVersion;
-import org.apache.hudi.stats.HoodieColumnRangeMetadata;
-import org.apache.hudi.stats.ValueMetadata;
+import org.apache.hudi.metadata.stats.HoodieColumnRangeMetadata;
+import org.apache.hudi.metadata.stats.ValueMetadata;
 
 import org.junit.jupiter.api.Test;
