+++ hudi-common/src/test/java/org/apache/hudi/metadata/stats/TestValueType.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.hudi.stats;
+package org.apache.hudi.metadata.stats;
 
 import org.junit.jupiter.api.Test;
